Complete task-01 epics 08-09

+++ erp-domain/src/main/java/com/bigbrightpaints/erp/modules/accounting/service/AccountingService.java
@@ -439,7 +439,7 @@ public class AccountingService {
         }
 
         if (foreignCurrency && totalForeignDebit.compareTo(BigDecimal.ZERO) > 0) {
-            entry.setForeignAmountTotal(totalForeignDebit.setScale(2, RoundingMode.HALF_UP).doubleValue());
+            entry.setForeignAmountTotal(totalForeignDebit.setScale(2, RoundingMode.HALF_UP));
         }
 
         if (duplicate.isPresent()) {
